Migrate register tests
