+++ src/qibo/tests/test_measurements.py
@@ -5,87 +5,6 @@ from qibo import gates, models
 from typing import Optional
 
 
-def test_register_name_error():
-    """Check that using the same register name twice results to error."""
-    c = models.Circuit(2)
-    c.add(gates.X(0))
-    c.add(gates.M(0, register_name="a"))
-    with pytest.raises(KeyError):
-        c.add(gates.M(1, register_name="a"))
-
-
-
-
-
-def test_measurement_qubit_order_multiple_registers(backend, accelerators):
-    c = models.Circuit(6, accelerators)
-    c.add(gates.X(0))
-    c.add(gates.X(1))
-    c.add(gates.X(3))
-    c.add(gates.M(5, 1, 3, register_name="a"))
-    c.add(gates.M(2, 0, register_name="b"))
-    result = c(nshots=100)
-
-    # Check full result
-    target_binary_samples = np.zeros((100, 5))
-    target_binary_samples[:, 1] = 1
-    target_binary_samples[:, 2] = 1
-    target_binary_samples[:, 4] = 1
-    assert_results(result,
-                   decimal_samples=13 * np.ones((100,)),
-                   binary_samples=target_binary_samples,
-                   decimal_frequencies={13: 100},
-                   binary_frequencies={"01101": 100})
-
-    target = {}
-    target["decimal_samples"] = {"a": 3 * np.ones((100,)),
-                                 "b": np.ones((100,))}
-    target["binary_samples"] = {"a": np.zeros((100, 3)),
-                                "b": np.zeros((100, 2))}
-    target["binary_samples"]["a"][:, 1] = 1
-    target["binary_samples"]["a"][:, 2] = 1
-    target["binary_samples"]["b"][:, 1] = 1
-
-    target["decimal_frequencies"] = {"a": {3: 100}, "b": {1: 100}}
-    target["binary_frequencies"] = {"a": {"011": 100}, "b": {"01": 100}}
-    assert_register_results(result, **target)
-
-def test_registers_in_circuit_with_unmeasured_qubits(accelerators):
-    """Check that register measurements are unaffected by unmeasured qubits."""
-    c = models.Circuit(5, accelerators)
-    c.add(gates.X(1))
-    c.add(gates.X(2))
-    c.add(gates.M(0, 2, register_name="A"))
-    c.add(gates.X(3))
-    c.add(gates.M(1, 4, register_name="B"))
-    result = c(nshots=100)
-
-    target = {}
-    target["decimal_samples"] = {"A": np.ones((100,)),
-                                 "B": 2 * np.ones((100,))}
-    target["binary_samples"] = {"A": np.zeros((100, 2)),
-                                "B": np.zeros((100, 2))}
-    target["binary_samples"]["A"][:, 1] = 1
-    target["binary_samples"]["B"][:, 0] = 1
-    target["decimal_frequencies"] = {"A": {1: 100}, "B": {2: 100}}
-    target["binary_frequencies"] = {"A": {"01": 100}, "B": {"10": 100}}
-    assert_register_results(result, **target)
-
-
-def test_registers_with_same_name_error():
-    """Check that registers with the same name cannot be added."""
-    c1 = models.Circuit(2)
-    c1.add(gates.H(0))
-    c1.add(gates.M(0))
-
-    c2 = models.Circuit(2)
-    c2.add(gates.H(1))
-    c2.add(gates.M(1))
-
-    with pytest.raises(KeyError):
-        c = c1 + c2
-
-
 def test_probabilistic_measurement(accelerators):
     import tensorflow as tf
     tf.random.set_seed(1234)
